oh wait

+++ src/shapes.c
@@ -122,9 +122,9 @@ void SR_DrawRect(
 void SR_DrawCircOutline(
     SR_Canvas *canvas,
     SR_RGBAPixel colour,
-    int x,
-    int y,
-    int r)
+    unsigned short x,
+    unsigned short y,
+    unsigned short r)
 {
     int xs = r;
     int ys = 0;
